% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffadjust.r
\name{diffadjust}
\alias{diffadjust}
\title{Adjusts difference}
\usage{
diffadjust(old.data, old.dataind, old.datainddiff)
}
\arguments{
\item{old.data}{}

\item{old.dataind}{}

\item{old.datainddiff}{}
}
\value{

}
\description{
Adjusts difference
}
\examples{

}
