% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHEPdata.r
\name{importHEPdata}
\alias{importHEPdata}
\title{Import HCV/HBV purchasing/dispensing data.}
\usage{
importHEPdata(filepath)
}
\arguments{
\item{filepath}{The location of the file you wish to import. File must be a 
.xls or .xlsx file. File must contain a single sheet.}
}
\value{
A data.frame
}
\description{
Import HCV/HBV purchasing/dispensing data.
}
\examples{

}
