% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2month.r
\name{num2month}
\alias{num2month}
\title{Return the Month name or Month name abbreviation}
\usage{
num2month(num, abbreviation)
}
\arguments{
\item{num}{A number from 1-12, representing a month.}

\item{abbreviation}{TRUE or FALSE, stating if you want the abbreviation.}
}
\value{
Either the full name of the month or the abbreviated name.
}
\description{
Return the Month name or Month name abbreviation
}
\examples{
num2month(2); will return 'February'
num2month(2,TRUE); will return 'Feb'. 
num2month(2,FALSE);  will return 'February' 

}
